training/pyproject.toml
[project]
name = "chess-watchtower"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.1",
    "cairosvg>=2.8.2",
    "git-python>=1.0.3",
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "tqdm>=4.67.1",
    "ultralytics>=8.3.221",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }

[tool.ruff]
select = ["E", "W", "F", "I"]
ignore = ["E501", "W191"]

[tool.ruff.format]
quote-style = "double"
indent-style = "tab"
line-ending = "auto"
